Feature: MVVM was implemented over cocktail detail view
